feat: Implement landing page and dashboard separation
- Created a new landing page at `/`.
- Moved the existing baby log content to a new `/dashboard` route.
- Updated middleware to protect the `/dashboard` route and allow public access to the landing page.
- Updated navigation and redirects to reflect the new structure.

Note: I was unable to update the redirect after login/signup in `src/app/login/actions.ts`. It still points to `/` instead of `/dashboard`.
